[polyLFO]
Add fractional frequency divisors, start of work on increasing
wavetable resolution

+++ software/o_c_REV/resources/frames_resources.py
@@ -79,5 +79,5 @@ resources = [
   (lookup_tables.lookup_tables_32,
    'lookup_table_hr', 'LUT', 'uint32_t', int, False),
   (waveforms.waveforms,
-   'wt', 'WT', 'uint8_t', int, False),
+   'wt', 'WT', 'uint16_t', int, False),
 ]
